+++ sdk/creditNote/index.js
@@ -1,9 +1,13 @@
 const creditNote = require('./creditNote.js');
 
 exports.log = require('./log');
+exports.Signer = require('./signer.js').Signer;
+exports.Invoice = require('./invoice.js').Invoice;
+exports.Transfer = require('./transfer.js').Transfer;
 exports.create = creditNote.create;
 exports.query = creditNote.query;
 exports.get = creditNote.get;
 exports.page = creditNote.page;
+exports.cancel = creditNote.cancel;
 exports.parse = creditNote.parse;
 exports.CreditNote = creditNote.CreditNote;