+++ sdk/webhook/index.js
@@ -1,0 +1,8 @@
+const webhook = require('./webhook.js');
+
+exports.create = webhook.create;
+exports.delete = webhook.delete;
+exports.query = webhook.query;
+exports.get = webhook.get;
+exports.page = webhook.page;
+exports.Webhook = webhook.Webhook;